requirements.txt
rudderlabs.data.apps
pyarrow==10.0.1
scikit-learn==1.0.2
pandas
pickle5